♻️ Refactor: 메인 페이지 로직, 스타일 분리
